Convolution 2D and Histogram Equalization/requirements.txt
opencv-python
numpy
matplotlib.pyplot